Move pdf.js to forms/shared
Fix lint issues.
