+++ server/test/routes/api/forms.test.js
@@ -7,7 +7,7 @@ import { authenticate, build } from '#test/helper.js';
 // Mock PDF generation before any imports reach the real #lib/pdf.js —
 // avoids needing Chromium in CI and keeps assertions focused on route wiring.
 // The actual renderToPdf rendering is an E2E / deployment smoke-test concern.
-mock.module('#lib/pdf.js', {
+mock.module('#lib/forms/shared/pdf.js', {
   namedExports: {
     renderFormToHtml: async () => '<html/>',
     renderToPdf: async () => Buffer.from('%PDF-mock'),
